+++ packages/koka-optic/CHANGELOG.md
@@ -1,5 +1,13 @@
 # koka-optic
 
+## 1.0.2
+
+### Patch Changes
+
+-   improve koka, add optic.proxy to koka-optic
+-   Updated dependencies
+    -   koka@1.0.3
+
 ## 1.0.1
 
 ### Patch Changes